move src/nearley_utils.ts to src/background/
