+++ src/background/nearley_utils.ts
